# payment-processing/src/main/java/com/payment/processing/dto/response/AuthResponse.java
package com.payment.processing.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class AuthResponse {
    private String accessToken;
    @Builder.Default
    private String tokenType = "Bearer";
    private Long expiresIn;
    private String userId;
    private String username;
    private String[] roles;
}